refactoring function structure
